+++ kifi-backend/modules/search/app/com/keepit/search/MainSearcher.scala
@@ -17,7 +17,6 @@ import com.keepit.search.graph.UserToUserEdgeSet
 import com.keepit.search.article.ArticleRecord
 import com.keepit.search.article.ArticleVisibility
 import com.keepit.search.index.Searcher
-import com.keepit.search.index.PersonalizedSearcher
 import com.keepit.search.spellcheck.SpellCorrector
 import com.keepit.search.query.HotDocSetFilter
 import com.keepit.search.query.QueryUtil
