+++ kifi-backend/modules/search/app/com/keepit/search/query/CollectionQuery.scala
@@ -4,7 +4,7 @@ import com.keepit.common.db.Id
 import com.keepit.common.logging.Logging
 import com.keepit.model.Collection
 import com.keepit.search.graph.CollectionToUriEdgeSet
-import com.keepit.search.index.PersonalizedSearcher
+import com.keepit.search.PersonalizedSearcher
 import org.apache.lucene.index.AtomicReaderContext
 import org.apache.lucene.index.IndexReader
 import org.apache.lucene.index.Term
