+++ kifi-backend/modules/search/app/com/keepit/search/query/SemanticVectorExtractorQuery.scala
@@ -2,7 +2,7 @@ package com.keepit.search.query
 
 import com.keepit.common.logging.Logging
 import com.keepit.search.SemanticVector
-import com.keepit.search.index.PersonalizedSearcher
+import com.keepit.search.PersonalizedSearcher
 import com.keepit.search.index.Searcher
 import org.apache.lucene.index.AtomicReaderContext
 import org.apache.lucene.index.IndexReader
