+++ kifi-backend/modules/search/test/com/keepit/search/query/SemanticVectorQueryTest.scala
@@ -7,6 +7,7 @@ import org.apache.lucene.util.Version
 import com.keepit.common.db.Id
 import com.keepit.common.db.SequenceNumber
 import com.keepit.search.index._
+import com.keepit.search.PersonalizedSearcher
 import com.keepit.search.SemanticVectorBuilder
 import java.io.StringReader
 import scala.Some
