+++ kifi-backend/modules/search/test/com/keepit/search/semantic/SemanticVarianceTest.scala
@@ -11,6 +11,8 @@ import com.keepit.search.query.TextQuery
 import org.apache.lucene.search.TermQuery
 import org.apache.lucene.index.Term
 import com.keepit.search.SemanticVectorBuilder
+import com.keepit.search.PersonalizedSearcher
+
 
 class SemanticVarianceTest extends Specification {
